backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

sqlmodel>=0.0.21
sqlalchemy>=2.0.0

python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

pydantic-settings>=2.3.0
python-dotenv>=1.0.0

livekit>=0.17.0
livekit-api>=0.7.0
livekit-agents>=0.9.0
livekit-plugins-google>=0.1.0

google-generativeai>=0.8.0

httpx>=0.27.0

redis[asyncio]>=5.0.0

boto3>=1.35.0

rtree>=1.2.0

numpy>=1.26.0
sentence-transformers>=3.0.0

#Utilities 
python-dateutil>=2.9.0
email-validator>=2.1.0
